lesson4: game code moved to GameScreen
