+++ android/src/ru/lazarev/game/AndroidLauncher.java
@@ -10,6 +10,6 @@ public class AndroidLauncher extends AndroidApplication {
 	protected void onCreate (Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
-		initialize(new myGame(), config);
+		initialize(new App(), config);
 	}
 }
