+++ desktop/src/ru/lazarev/game/DesktopLauncher.java
@@ -8,6 +8,6 @@ public class DesktopLauncher {
 	public static void main (String[] arg) {
 		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
 		config.setForegroundFPS(60);
-		new Lwjgl3Application(new myGame(), config);
+		new Lwjgl3Application(new App(), config);
 	}
 }
